requirements.txt
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.0
